requirements.txt
# Django framework -> https://www.djangoproject.com
Django==5.1.4

# Django REST Framework -> https://www.django-rest-framework.org
djangorestframework==3.14.0

# drf-yasg -> https://github.com/axnsan12/drf-yasg
drf-yasg==1.21.7

# requests -> https://pypi.org/project/requests/
requests==2.32.3

# setuptools https://pypi.org/project/setuptools/ 
setuptools==75.6.0 

